requirements.txt
langchain~=0.3.26
python-dotenv~=1.1.1
google-api-python-client
faiss-cpu~=1.11.0.post1
chainlit~=2.6.2
python-dotenv
requests~=2.32.4

protobuf~=5.29.5
langchain-core~=0.3.70
langchain-google-genai~=2.0.10
langchain-community~=0.3.27
pydantic~=2.11.7